+++ CHANGELOG.md
@@ -8,3 +8,4 @@ HEAD
 * Add Release build badge on Readme
 * Add version to PlatformIO Registry badge
 * Add PayPal donate badge
+* Replace WebConfigMQTT with MQTTClient and remove PubSubClient dependencies